Create CalendarCalculator and move methods from CalndarConnector
